+++ AMWP/Models/Chart.cs
@@ -8,7 +8,7 @@ namespace AMWP.Models
     public class Chart
     {
         public List<string> labels { get; set; }
-        public List<decimal> data { get; set; }
+        public List<double> data { get; set; }
     }
     
 }
